# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 智慧云脑诊疗平台 API
  version: 0.1.0
  description: |
    阶段二：基础数据管理（认证 + 科室 + 患者 + 医生 + 药品）
    阶段三：排班与挂号（Phase 5-6）
    阶段四：分诊与就诊（Phase 7-8）

    当前包含接口：
    - 认证模块：login / logout / change-password（3 个）
    - 科室模块：tree / list / create / update（4 个）
    - 患者模块：register / me / detail / update / profile / search（6 个）
    - 医生模块：list / create / update / by-department（4 个）
    - 药品模块：list / detail / by-code / by-category / interaction / contraindications（7 个）
    - 排班模块：create / update / cancel / detail / by-doctor / by-department / available（7 个）
    - 挂号模块：create / cancel / detail / by-patient / by-doctor / pending / all（7 个）
    - 分诊模块：analyze / detail / by-patient（3 个）
    - 就诊模块：start / wait-exam / resume / complete / cancel / diagnoses / detail（12 个）

    契约来源：contracts/30_接口数据与错误契约.md 第11节 + 第12节
servers:
  - url: http://localhost:8080
    description: 本地后端
  - url: http://localhost:4010
    description: Prism API Mock Server
paths:
...
  /api/triage/analyze:
    post:
      tags:
        - 分诊
      summary: AI 分诊分析
      description: |
        患者输入症状，返回 AI 分诊建议、科室映射和推荐可预约排班。
        AI 失败时返回降级标记，提示转人工选择科室。
      operationId: analyzeTriage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/triage_encounter.yaml#/TriageAnalyzeRequest'
      responses:
        '200':
          description: 分诊分析结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/TriageAnalyzeResponse'
        '403':
          description: 患者不活跃或无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 患者不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/triage/{id}:
    get:
      tags:
        - 分诊
      summary: 获取分诊记录详情
      operationId: getTriageRecordById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 分诊记录详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/TriageRecordResponse'
        '404':
          description: 分诊记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/triage/patient/{patientId}:
    get:
      tags:
        - 分诊
      summary: 查询患者分诊记录列表（分页）
      operationId: getTriageRecordsByPatient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 分诊记录列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================
  # 就诊接口（阶段四）
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      $ref: ./schemas/common.yaml#/ApiResponse
    PageResponse:
      $ref: ./schemas/common.yaml#/PageResponse
...